\name{ann}
\alias{ann}
\docType{data}
\title{
Annotations.
}
\description{
Annotation for example dataset.
}
\usage{data("ann")}
\format{
  A data frame with 73 observations on the following 5 variables.
  \describe{
    \item{\code{Chir1Data}}{a character vector, with thendate of 1st surgical event.}
    \item{\code{UltimoFUPData}}{the date of the last followup.}
    \item{\samp{0}}{status}
    \item{\code{Grado}}{Degree}
    \item{\code{Isto}}{Histotype.}
  }
}
\examples{
data(ann)
}
\keyword{datasets}
