% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survClip-backbone.R
\name{computeDays}
\alias{computeDays}
\title{Function to compute days}
\usage{
computeDays(timeTable)
}
\arguments{
\item{timeTable}{a table with starting date and ending date}
}
\value{
the days
}
\description{
A simple function to start from an interval of dates to days
}
