\name{exp}
\alias{exp}
\docType{data}
\title{
Expression dataset.
}
\description{
Expression dataset for survival test.
}
\usage{data("exp")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(Exp)
}
\keyword{datasets}
