% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{graph}
\alias{graph}
\title{An example graphNEL}
\format{A graphNEL object of a pathway with 308 nodes.}
\usage{
graph
}
\description{
Toy example of graphNEL object for pathway the survival test
}
\keyword{datasets}
