% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{survAnnot}
\alias{survAnnot}
\title{Anotations}
\format{An object of class \code{data.frame} with 73 rows and 2 columns.}
\usage{
survAnnot
}
\description{
Survival annotation for toy example dataset
}
\details{
#' @format A data frame with 73 observations on the following 2 variables.
\describe{
  \item{\code{days}}{a character vector, with the days to Death/last follow up after 1st surgical event.}
  \item{\code{status}}{dead = 1, alive = 0}
}
}
\keyword{datasets}
