% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-class.R
\docType{class}
\name{survCliques-class}
\alias{survCliques-class}
\alias{survCliques}
\alias{survPath-class}
\alias{survPath}
\title{Survival Cliques Class}
\description{
Class to describe and represent survival analysis on modules

Class to describe and represent survival analysis on modules
}
\details{
For internal use only

For internal use only
}
\section{Slots}{

\describe{
\item{\code{alphas}}{Object of class \code{"numeric"}. It contains pvalues of the cliques/modules}

\item{\code{zlist}}{Object of class \code{"list"}. For each cliques, the list of pvalues for all the covariates}

\item{\code{cliques}}{Object of class \code{"list"}. The list of the genes in the cliques}

\item{\code{coxObjs}}{Object of class \code{"list"}. For each cliques, the object used for Coxph analysis}

\item{\code{cliquesLoadings}}{Object of class \code{"list"}. For each cliques, the loadings as calculated by the PCA}

\item{\code{cliquesExpr}}{Object of class \code{"list"}. For each cliques, the cliques expression}

\item{\code{pvalue}}{Object of class \code{"numeric"}. It contains the pvalue of the whole model generated by the analysis}

\item{\code{zlist}}{Object of class \code{"numeric"}. vector of pvalues for all the covariates}

\item{\code{coxObj}}{Object of class \code{"data.frame"}. The object used for Coxph analysis}

\item{\code{loadings}}{Object of class \code{"matrix"}. The loadings as calculated by the PCA}

\item{\code{method}}{Object of class \code{"character"}. Type of method used}
}}

\examples{
showClass("survCliques")

showClass("survPath")

}
