% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survClip-backbone.R
\name{survivalcoxr}
\alias{survivalcoxr}
\alias{coxrsummary}
\title{Cox Robust Model Analysis}
\usage{
survivalcoxr(coxObj, formula)

coxrsummary(x)
}
\arguments{
\item{coxObj}{data.frame: patients x covariates}

\item{formula}{formula to use}

\item{x}{a coxr.obj}
}
\value{
A list with
\item{pvalue}{pvalue of the model}
\item{zlist}{pvalues of single covariates}
\item{coxObj}{the original coxObj passed to the function}

a list with wald test and robust and partial coefficients
}
\description{
Cox Robust Analysis
}
\details{
For internal use only
}
