% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqueSurvivalTest.R
\name{cliqueSurvivalTest}
\alias{cliqueSurvivalTest}
\title{Perform survival test on all the cliques of the graph}
\usage{
cliqueSurvivalTest(expr, survAnnot, graph,
  pcsSurvCoxMethod = c("regular", "sparse"), alwaysShrink = FALSE,
  maxPCs = 10, survFormula = "Surv(days, status) ~", robust = FALSE)
}
\arguments{
\item{expr}{expression matrix}

\item{survAnnot}{a data frame for survival annotations specified according to the survFormula. The data frame must contain days and status}

\item{graph}{a graphNEL object for a graph}

\item{pcsSurvCoxMethod}{a method to perform PCA. Can be "regular", "topological", "sparse" for regular PCA, topological based PCA and sparse PCA, respectively. The latter one (sparse) is particularly suited for cliques only}

\item{alwaysShrink}{if TRUE, always shrink the covariance matrix. Deafult=FALSE}

\item{maxPCs}{maximum number of PCs used in the cox formula "Surv(days, status) ~ PC1.."}

\item{survFormula}{the formula used in Coxph analysis. Defaut="Surv(days, status) ~". Please note that the formula end with '~' meaning that PCs will be added}

\item{robust}{should be used the robust mode for cox}
}
\value{
A survCliques object
}
\description{
This function performs survival test on given pathway using a matrix with survival annotation
}
\details{
Survival test is made according to survFormula. With "regular" method, a regular PCA analysis is used to compute PCs. With "sparse" method, a penalized regression is used for the estimation of PCs (as implemented in elasticnet)
}
\examples{
if (require(graphite)) {
  data(exp)
  data(survAnnot)
  data(graph)
  row.names(exp) <- paste0("ENTREZID:", row.names(exp))
  genes <- intersect(graph::nodes(graph), row.names(exp))
  graph <- graph::subGraph(genes, graph)
  expr <- exp[genes, , drop=FALSE]
  cliqueSurvivalTest(expr, survAnnot, graph, maxPCs=2)
}

}
\seealso{
\code{\link{pathwaySurvivalTest}}, \code{\link{getTopLoadGenes}}
}
