\name{computeDays}
\alias{computeDays}

\title{
Function to compute days.
}
\description{
A simple function to start from an interval of dates to days.
}
\usage{
computeDays(timeTable)
}
\arguments{
  \item{timeTable}{a table with starting date and ending date.}
}
\value{the days}
\author{Paolo Martini}
