% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exp}
\alias{exp}
\title{Expression dataset}
\format{A matrix with 246 genes (rows) measured across 73 patients (columns)}
\usage{
exp
}
\description{
Toy example of an expression dataset for survival test
}
\keyword{datasets}
