\name{survAnnot}
\alias{survAnnot}
\docType{data}
\title{
Annotations.
}
\description{
Survival annotation for toy example dataset.
}
\usage{data("survAnnot")}
\format{
  A data frame with 73 observations on the following 2 variables.
  \describe{
    \item{\code{days}}{a character vector, with the days to Death/last follow up after 1st surgical event.}
    \item{\code{status}}{dead = 1, alive = 0}
  }
}
\examples{
data(survAnnot)
}
\keyword{datasets}
