% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survClip-backbone.R
\name{survivalcox}
\alias{survivalcox}
\title{Cox Model Analysis}
\usage{
survivalcox(coxObj, formula)
}
\arguments{
\item{coxObj}{data.frame: patients x covariates}

\item{formula}{formula to use}
}
\value{
A list with
\item{pvalue}{pvalue of the model}
\item{zlist}{pvalues of single covariates}
\item{coxObj}{the original coxObj passed to the function}
}
\description{
Cox Analysis
}
\details{
For internal use only
}
