\name{survPath-class}
\docType{class}
\alias{survPath-class}
\alias{survPath}

\title{Class \code{"survPath"}}
\description{Descriptor for survival pathway analysis.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("survPath", ...)}.
pvalues and the method used
}
\section{Slots}{
  \describe{
    \item{\code{.Data}:}{Object of class \code{"list"} }
    \item{\code{pvalues}:}{Object of class \code{"list"} pvalues generated by the analysis }
    \item{\code{method}:}{Object of class \code{"character"} type of method used. }
  }
}
\section{Extends}{
Class \code{"\linkS4class{list}"}, from data part.
Class \code{"\linkS4class{vector}"}, by class "list", distance 2.
Class \code{"\linkS4class{Ulist}"}, by class "list", distance 2.
Class \code{"\linkS4class{Uvector}"}, by class "list", distance 3.
}
\section{Methods}{
No methods defined with class "survPath" in the signature.
}
\author{
Paolo Martini
}
\examples{
showClass("survPath")
}
\keyword{classes}
